CRUD для NPC;

+++ BRIX.Mobile/ViewModel/NPCs/NPCsPageVM.cs
@@ -1,9 +1,12 @@
 ﻿using BRIX.Library.Characters;
 using BRIX.Mobile.Models.NPCs;
+using BRIX.Mobile.Resources.Localizations;
 using BRIX.Mobile.Services;
 using BRIX.Mobile.Services.Navigation;
 using BRIX.Mobile.View.NPCs;
 using BRIX.Mobile.ViewModel.Base;
+using BRIX.Mobile.ViewModel.Popups;
+using BRIX.Utility.Extensions;
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
 
@@ -34,35 +37,55 @@ namespace BRIX.Mobile.ViewModel.NPCs
         {
             await Navigation.NavigateAsync<AOENPCsPage>(
                 (NavigationParameters.EditMode, EEditingMode.Edit),
-                (NavigationParameters.NPC, EEditingMode.Edit)
+                (NavigationParameters.NPC, item.Copy())
             );
         }
 
         [RelayCommand]
-        public async Task RemmoveNPC(NPCModel item)
+        public async Task RemoveNPC(NPCModel item)
         {
-            await _characterService.RemoveNPC(item.Internal);
+            AlertPopupResult? result = await Ask(Localization.RemoveNPCQuestion);
+
+            if (result?.Answer == EAlertPopupResult.Yes)
+            {
+                await _characterService.RemoveNPC(item.Internal);
+                NPCs.Remove(item);
+            }
         }
 
-        public void ApplyQueryAttributes(IDictionary<string, object> query)
+        public async void ApplyQueryAttributes(IDictionary<string, object> query)
         {
-            HandleBackFormEditing(query);
+            await HandleBackFormEditing(query);
+            query.Clear();
         }
 
-        private void HandleBackFormEditing(IDictionary<string, object> query)
+        private async Task HandleBackFormEditing(IDictionary<string, object> query)
         {
-            
+            NPCModel? npc = query.GetParameterOrDefault<NPCModel>(NavigationParameters.NPC);
+
+            if (npc != null)
+            {
+                EEditingMode mode = query.GetParameterOrDefault<EEditingMode>(NavigationParameters.EditMode);
+
+                switch (mode)
+                {
+                    case EEditingMode.Add:
+                        await _characterService.AddNPC(npc.Internal);
+                        NPCs.Add(npc);
+                        break;
+                    case EEditingMode.Edit:
+                        await _characterService.UpdateNPC(npc.Internal);
+                        int index = NPCs.IndexOf(NPCs.First(x => x.Internal.Id == npc.Internal.Id));
+                        NPCs[index] = npc;
+                        break;
+                }
+            }
         }
 
         public override async Task OnNavigatedAsync()
         {
             List<NPC> npcs = await _characterService.GetNPCs();
             NPCs = new(npcs.Select(x => new NPCModel(x)));
-            //{
-            //    new NPCItemVM { Name = "Мусорный гоблин", Power = 70 },
-            //    new NPCItemVM { Name = "Дьявольская гидра", Power = 66666 },
-            //    new NPCItemVM { Name = "Крыса по имени Генри", Power = 5 },
-            //};
         }
     }
 }
